refactor: CouponResponseItem 제거 및 CouponResponse 구조 변경
